+++ src/samplers/python/BaseSamplers.cpp
@@ -48,6 +48,8 @@
 #include <utk/samplers/SamplerBridson.hpp>
 #include <utk/samplers/SamplerProjectiveBlueNoise.hpp>
 #include <utk/samplers/SamplerSobol.hpp>
+#include <utk/samplers/SamplerCascadedSobol.hpp>
+#include <utk/samplers/SamplerGenerativeMatrices.hpp>
 #include <utk/samplers/SamplerR1.hpp>
 #include <utk/samplers/SamplerKronecker.hpp>
 #include <utk/samplers/SamplerKorobov.hpp>
@@ -328,4 +330,30 @@ void init_BaseSampler(py::module& m)
         .def("setSeed", GetSetSeedFunction<SobolSampler>()              , py::arg("seed") = NO_SEED)
         .def("sample",  GetSampleFunction <SobolSampler>()              , py::arg("N"))
         .def("isample", GetSampleFunction <SobolSampler, uint32_t>(), py::arg("N"));
+
+    
+    using CascadedSobolSampler = SamplerCascadedSobol<uint32_t>; 
+    py::class_<CascadedSobolSampler>(m, "CascadedSobol")
+        .def(
+            py::init<uint32_t, uint32_t>(),
+            py::arg("d"), py::arg("depth") = 32
+        )
+        .def("__repr__", [](const CascadedSobolSampler& wn) { return "CascadedSobolSampler(d=" + std::to_string(wn.GetDimension()) +")"; })
+        .def("setDirectionFile", &CascadedSobolSampler::setDirectionFile)
+        .def("setOwenDepth",     &CascadedSobolSampler::setOwenDepth)
+        .def("setSeed", GetSetSeedFunction<CascadedSobolSampler>()          , py::arg("seed") = NO_SEED)
+        .def("sample",  GetSampleFunction <CascadedSobolSampler>()          , py::arg("N"))
+        .def("isample", GetSampleFunction <CascadedSobolSampler, uint32_t>(), py::arg("N"));
+
+    using GenerativeMatrices = SamplerGenerativeMatrices<uint32_t>; 
+    py::class_<GenerativeMatrices>(m, "GenerativeMatrices")
+        .def(
+            py::init<std::string, uint8_t, uint8_t, uint8_t, bool>(),
+            py::arg("filename"), py::arg("m"), py::arg("p"), py::arg("d"), py::arg("scramble") = false
+        )
+        .def("__repr__", [](const GenerativeMatrices& wn) { return "GenerativeMatricesSampler(d=" + std::to_string(wn.GetDimension()) +")"; })
+        .def("setSeed", GetSetSeedFunction<GenerativeMatrices>()          , py::arg("seed") = NO_SEED)
+        .def("sample",  GetSampleFunction <GenerativeMatrices>()          , py::arg("N"))
+        .def("isample", GetSampleFunction <GenerativeMatrices, uint32_t>(), py::arg("N"));
+
 }